[FULLSTACK] Allow decimal values on nutritional stats

+++ efw-back/src/main/java/com/eliasfb/efw/dto/CreateOrUpdateIngredientDto.java
@@ -11,9 +11,9 @@ public class CreateOrUpdateIngredientDto implements Serializable {
 
 	private Integer userId;
 	private String name;
-	private Integer calories;
-	private Integer proteins;
-	private Integer fats;
-	private Integer carbohydrates;
+	private Double calories;
+	private Double proteins;
+	private Double fats;
+	private Double carbohydrates;
 
 }
